requirements.txt
torch>=2.2
chromadb>=0.5
sentence-transformers>=3.0
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
scipy>=1.11
pyyaml>=6.0
tqdm>=4.66
